LayoutTests/svg/animations/multiple-animations-ending.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../dynamic-updates/resources/SVGTestCase.js"></script>
<script src="resources/SVGAnimationTestCase.js"></script>
</head>
<body onload="runSMILTest()">
<h1>Test the effect of multiple animations ending.</h1>
<p id="description"></p>
<div id="console"></div>
<script src="script-tests/multiple-animations-ending.js"></script>
</body>
</html>
