LayoutTests/svg/animations/multiple-animations-ending-expected.txt

Test the effect of multiple animations ending.


This checks the effect on multiple animations ending on one target

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 100
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 200
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 10
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 80
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 210
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 30
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 80
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 210
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 30
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 80
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 210
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 30
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 60
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 220
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 30
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 0
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 60
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 0
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 60
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 5
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 80
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 200
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 5
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 80
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 200
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 5
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 80
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 225
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 10
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 80
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 225
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 20
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 110
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 225
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 20
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 110
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 130
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 130
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 130
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 5
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 130
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 250
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 160
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 250
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 160
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 200
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 180
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 200
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 180
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 150
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 180
PASS rect4.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 200
PASS rect3.x.animVal.value is 150
PASS rect3.x.baseVal.value is 0
PASS rect4.x.animVal.value is 180
PASS rect4.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
